@@ -1,6 +1,6 @@
 module github.com/cline/cli
 
-go 1.23.0
+go 1.24.0
 
 require (
 	github.com/atotto/clipboard v0.1.4
