@@ -47,7 +47,7 @@ func GetProviderConfigurations(ctx context.Context, manager *task.Manager) (*Pro
 	}
 
 	// Parse state_json as map[string]interface{}
-	var stateData map[string]interface{}
+	var stateData map[string]any
 	if err := json.Unmarshal([]byte(stateJSON), &stateData); err != nil {
 		return nil, fmt.Errorf("failed to parse state JSON: %w", err)
 	}
@@ -57,7 +57,7 @@ func GetProviderConfigurations(ctx context.Context, manager *task.Manager) (*Pro
 	}
 
 	// Extract apiConfiguration object from state
-	apiConfig, ok := stateData["apiConfiguration"].(map[string]interface{})
+	apiConfig, ok := stateData["apiConfiguration"].(map[string]any)
 	if !ok {
 		if global.Config.Verbose {
 			fmt.Println("[DEBUG] No apiConfiguration found in state")
@@ -128,11 +128,11 @@ func (r *ProviderListResult) GetAllReadyProviders() []*ProviderDisplay {
 		modelID := getProviderSpecificModelID(r.apiConfig, "plan", provider)
 
 		// Determine if credentials exist
-		hasCreds := checkAPIKeyExists(r.apiConfig, provider)
+		hasCreds := checkCredentialsExists(r.apiConfig, provider)
 
 		// Determine readiness: OCA uses auth state presence; others need creds and model
 		if provider == cline.ApiProvider_OCA {
-			state, _ := GetLatestOCAState(context.Background(), 2 *time.Second)
+			state, _ := GetLatestOCAState(context.Background(), 2*time.Second)
 			if state == nil || state.User == nil {
 				continue
 			}
@@ -156,7 +156,7 @@ func (r *ProviderListResult) GetAllReadyProviders() []*ProviderDisplay {
 			Mode:      "Ready",
 			Provider:  provider,
 			ModelID:   modelID,
-			HasAPIKey: checkAPIKeyExists(r.apiConfig, provider),
+			HasAPIKey: checkCredentialsExists(r.apiConfig, provider),
 			BaseURL:   baseURL,
 		})
 		seenProviders[provider] = true
@@ -192,7 +192,7 @@ func extractProviderFromState(stateData map[string]interface{}, mode string) *Pr
 	modelID := getProviderSpecificModelID(stateData, mode, provider)
 
 	// Check if API key exists
-	hasAPIKey := checkAPIKeyExists(stateData, provider)
+	hasCredentials := checkCredentialsExists(stateData, provider)
 
 	// Get base URL for Ollama (can be shown publicly)
 	baseURL := ""
@@ -206,7 +206,7 @@ func extractProviderFromState(stateData map[string]interface{}, mode string) *Pr
 		Mode:      capitalizeMode(mode),
 		Provider:  provider,
 		ModelID:   modelID,
-		HasAPIKey: hasAPIKey,
+		HasAPIKey: hasCredentials,
 		BaseURL:   baseURL,
 	}
 }
@@ -215,7 +215,7 @@ func extractProviderFromState(stateData map[string]interface{}, mode string) *Pr
 // Returns (provider, ok) where ok is false if the provider is unknown
 func mapProviderStringToEnum(providerStr string) (cline.ApiProvider, bool) {
 	normalizedStr := strings.ToLower(providerStr)
-	
+
 	// Map string values to enum values
 	switch normalizedStr {
 	case "anthropic":
@@ -303,19 +303,23 @@ func getProviderSpecificModelID(stateData map[string]interface{}, mode string, p
 	return modelID
 }
 
-// checkAPIKeyExists checks if API key field exists in state (never retrieve actual key)
-func checkAPIKeyExists(stateData map[string]interface{}, provider cline.ApiProvider) bool {
+// checkCredentialsExists checks if API key field exists in state (never retrieve actual key)
+func checkCredentialsExists(stateData map[string]interface{}, provider cline.ApiProvider) bool {
 	// Get field mapping from centralized function
 	fields, err := GetProviderFields(provider)
 	if err != nil {
 		return false
 	}
 
-	keyField := fields.APIKeyField
-
 	// Check if the key exists and is not empty
-	if value, ok := stateData[keyField]; ok {
+	if value, ok := stateData[fields.APIKeyField]; ok {
 		if str, ok := value.(string); ok && str != "" {
+			return true
+		}
+	}
+
+	if value, ok := stateData[fields.UseProfileField]; ok {
+		if hasProfileField, ok := value.(bool); ok && hasProfileField {
 			return true
 		}
 	}
@@ -438,13 +442,13 @@ func DetectAllConfiguredProviders(ctx context.Context, manager *task.Manager) ([
 	stateJSON := state.StateJson
 
 	// Parse state_json as map[string]interface{}
-	var stateData map[string]interface{}
+	var stateData map[string]any
 	if err := json.Unmarshal([]byte(stateJSON), &stateData); err != nil {
 		return nil, fmt.Errorf("failed to parse state JSON: %w", err)
 	}
 
 	// Extract apiConfiguration object from state
-	apiConfig, ok := stateData["apiConfiguration"].(map[string]interface{})
+	apiConfig, ok := stateData["apiConfiguration"].(map[string]any)
 	if !ok {
 		verboseLog("[DEBUG] No apiConfiguration found in state")
 		verboseLog("[DEBUG] Available keys in stateData: %v", getMapKeys(stateData))
@@ -469,35 +473,37 @@ func DetectAllConfiguredProviders(ctx context.Context, manager *task.Manager) ([
 
 	// Check each BYO provider for API key presence
 	providersToCheck := []struct {
-		provider cline.ApiProvider
-		keyField string
+		provider  cline.ApiProvider
+		keyFields []string
 	}{
-		{cline.ApiProvider_ANTHROPIC, "apiKey"},
-		{cline.ApiProvider_OPENAI, "openAiApiKey"},
-		{cline.ApiProvider_OPENAI_NATIVE, "openAiNativeApiKey"},
-		{cline.ApiProvider_OPENROUTER, "openRouterApiKey"},
-		{cline.ApiProvider_XAI, "xaiApiKey"},
-		{cline.ApiProvider_BEDROCK, "awsAccessKey"},
-		{cline.ApiProvider_GEMINI, "geminiApiKey"},
-		{cline.ApiProvider_OLLAMA, "ollamaBaseUrl"}, // Ollama uses baseUrl instead of API key
-		{cline.ApiProvider_CEREBRAS, "cerebrasApiKey"},
-		{cline.ApiProvider_HICAP, "hicapApiKey"},
-		{cline.ApiProvider_NOUSRESEARCH, "nousResearchApiKey"},
+		{cline.ApiProvider_ANTHROPIC, []string{"apiKey"}},
+		{cline.ApiProvider_OPENAI, []string{"openAiApiKey"}},
+		{cline.ApiProvider_OPENAI_NATIVE, []string{"openAiNativeApiKey"}},
+		{cline.ApiProvider_OPENROUTER, []string{"openRouterApiKey"}},
+		{cline.ApiProvider_XAI, []string{"xaiApiKey"}},
+		{cline.ApiProvider_BEDROCK, []string{"awsAccessKey", "awsUseProfile"}},
+		{cline.ApiProvider_GEMINI, []string{"geminiApiKey"}},
+		{cline.ApiProvider_OLLAMA, []string{"ollamaBaseUrl"}}, // Ollama uses baseUrl instead of API key
+		{cline.ApiProvider_CEREBRAS, []string{"cerebrasApiKey"}},
+		{cline.ApiProvider_HICAP, []string{"hicapApiKey"}},
+		{cline.ApiProvider_NOUSRESEARCH, []string{"nousResearchApiKey"}},
 	}
 
 	for _, providerCheck := range providersToCheck {
-		verboseLog("[DEBUG] Checking for %s key: %s", GetProviderDisplayName(providerCheck.provider), providerCheck.keyField)
-		if value, ok := apiConfig[providerCheck.keyField]; ok {
-			verboseLog("[DEBUG]   Found key, value type: %T, is empty: %v", value, value == "")
-			if str, ok := value.(string); ok && str != "" {
-				configuredProviders = append(configuredProviders, providerCheck.provider)
-				verboseLog("[DEBUG]   ✓ Provider %s is configured", GetProviderDisplayName(providerCheck.provider))
+		verboseLog("[DEBUG] Checking for %s key: %s", GetProviderDisplayName(providerCheck.provider), providerCheck.keyFields)
+		for _, keyField := range providerCheck.keyFields {
+			if value, ok := apiConfig[keyField]; ok {
+				verboseLog("[DEBUG]   Found key, value type: %T, is empty: %v", value, value == "")
+				if str, ok := value.(string); ok && str != "" {
+					configuredProviders = append(configuredProviders, providerCheck.provider)
+					verboseLog("[DEBUG]   ✓ Provider %s is configured", GetProviderDisplayName(providerCheck.provider))
+					break
+				}
+			} else {
+				verboseLog("[DEBUG]   Key %s not found", keyField)
 			}
-		} else {
-			verboseLog("[DEBUG]   Key %s not found", providerCheck.keyField)
 		}
 	}
-
 
 	verboseLog("[DEBUG] Total configured providers: %d", len(configuredProviders))
 	for _, p := range configuredProviders {
