@@ -8,6 +8,10 @@
  * @returns true if hooks are enabled and supported on this platform, false otherwise
  */
 export function getHooksEnabledSafe(userSetting: boolean | undefined): boolean {
+	// Handle legacy object format: {user: boolean, featureFlag: boolean}, which
+	// can occur if the migration hasn't run yet or if reading from an old state.
+	const booleanValue = Boolean((userSetting as any)?.user ?? userSetting)
+
 	// Force hooks to false on Windows (not yet supported)
-	return process.platform === "win32" ? false : (userSetting ?? false)
+	return process.platform === "win32" ? false : booleanValue
 }
