@@ -129,7 +129,7 @@ const GEMINI_3_RULES_TEMPLATE = (_context: SystemPromptContext) => `RULES
 
 - The current working directory is \`{{CWD}}\` - this is the directory where all the tools will be executed from.
 - When executing terminal commands, new terminals always open in the workspace directory. Use relative paths or chain commands with proper shell operators (e.g., \`cd path && command\` to change directory and run a command together).
-- Whean searching, prefer the search_files tool over using grep in the terminal. If you are directly instruted to use grep, ensure your search patterns are targetted and not too vague to prevent extremely large outputs.
+- When searching, prefer the search_files tool over using grep in the terminal. If you are directly instructed to use grep, ensure your search patterns are targeted and not too vague to prevent extremely large outputs.
 - When using replace_in_file, pay careful attention to the EDITING FILES section above. The most common errors are:
   - Not matching content exactly (every character, space, and newline must match)
   - Using incomplete lines in SEARCH blocks (always include complete lines from start to end)
@@ -157,13 +157,13 @@ Plan Mode is for deep analysis and strategic planning before implementation. You
 
 ### Phase 1: Silent Investigation
 
-Perform comprehensive research to build complete understanding of the codebase. Work silently - execute targetted searcg commands and read files without explaining what you're doing. Only ask questions when truly necessary for planning. You must strongly incrporate key words and principles from the user's input into your targetted search patterns and strategy.
+Perform comprehensive research to build complete understanding of the codebase. Work silently - execute targeted search commands and read files without explaining what you're doing. Only ask questions when truly necessary for planning. You must strongly incorporate key words and principles from the user's input into your targeted search patterns and strategy.
 
 **Research Activities:**
 - Use read_file, search_files, and list_code_definition_names extensively to understand architecture, patterns, and conventions
-- Execute targetted terminal commands to search and gather information about structure and dependencies.
+- Execute targeted terminal commands to search and gather information about structure and dependencies.
 - Identify technical constraints, existing patterns, and potential risks${context.yoloModeToggled !== true ? "\n- Ask targeted clarifying questions only when they will directly influence your implementation approach" : ""}
-- Ensure complete converage- before presenting a plan, you should identify all related functions, classes, calls, and methods that are involved or affected by the proposed changes.
+- Ensure complete coverage - before presenting a plan, you should identify all related functions, classes, calls, and methods that are involved or affected by the proposed changes.
 
 ### Phase 2: Plan Presentation
 
@@ -198,7 +198,7 @@ Engage with the user to discuss the plan, answer questions, and incorporate feed
 
 ### Phase 4: Transition to Implementation
 
-Once the plan is finalized and approved, you MUST direct the user to switch to ACT MODE. In Act Mode, you'll execute the plan step-by-step as outlined. If you not specifically ask the user to switch to ACT MODE, you will not be able to implemnent the planned changes.
+Once the plan is finalized and approved, you MUST direct the user to switch to ACT MODE. In Act Mode, you'll execute the plan step-by-step as outlined. If you not specifically ask the user to switch to ACT MODE, you will not be able to implement the planned changes.
 
 ## Act Mode Workflow
 
