@@ -228,7 +228,7 @@ function createAuthSucceededHtml(redirectUri?: string): string {
     <title>Cline - Authentication Success</title>
 	${redirect}
     <style>
-        @import url('https://fonts.googleapis.com/css2?family=Azeret+Mono:wght@300;400;700&display=swap');
+        @import url('https://fonts.googleapis.com/css2?family=Azeret:wght@300;400;700&display=swap');
         
         * {
             margin: 0;
@@ -237,7 +237,7 @@ function createAuthSucceededHtml(redirectUri?: string): string {
         }
         
         body {
-            font-family: 'Azeret Mono', monospace;
+            font-family: 'Azeret', sans-serif;
             background-color: #ffffff;
             color: #333333;
             height: 100vh;
