@@ -126,13 +126,13 @@ describe("BannerService", () => {
 	})
 
 	describe("API Provider Rule Evaluation (Client-Side)", () => {
-		it("should show banner when user has the required API provider configured", async () => {
+		it("should show banner when user has selected the required API provider in act mode", async () => {
 			const controllerWithOpenAI: Partial<Controller> = {
 				stateManager: {
 					getApiConfiguration: () => ({
-						openAiApiKey: "sk-test-key",
+						actModeApiProvider: "openai",
 					}),
-					getGlobalSettingsKey: () => undefined,
+					getGlobalSettingsKey: (key: string) => (key === "mode" ? "act" : undefined),
 					getGlobalStateKey: () => [],
 				} as any,
 			}
@@ -164,19 +164,57 @@ describe("BannerService", () => {
 			expect(banners[0].id).to.equal("bnr_openai")
 		})
 
-		it("should NOT show banner when user doesn't have the required API provider", async () => {
-			const controllerWithoutOpenAI: Partial<Controller> = {
+		it("should show banner when user has selected the required API provider in plan mode", async () => {
+			const controllerWithAnthropic: Partial<Controller> = {
 				stateManager: {
 					getApiConfiguration: () => ({
-						apiKey: "sk-ant-test", // Has Anthropic key but not OpenAI
+						planModeApiProvider: "anthropic",
 					}),
-					getGlobalSettingsKey: () => undefined,
+					getGlobalSettingsKey: (key: string) => (key === "mode" ? "plan" : undefined),
 					getGlobalStateKey: () => [],
 				} as any,
 			}
 			// Reinitialize with new controller
 			BannerService.reset()
-			bannerService = BannerService.initialize(controllerWithoutOpenAI as Controller)
+			bannerService = BannerService.initialize(controllerWithAnthropic as Controller)
+
+			const mockResponse = {
+				data: {
+					data: {
+						items: [
+							{
+								id: "bnr_anthropic",
+								titleMd: "Anthropic Users",
+								bodyMd: "For Anthropic API",
+								severity: "info" as const,
+								placement: "top" as const,
+								rulesJson: JSON.stringify({ providers: ["anthropic"] } as BannerRules),
+							},
+						],
+					},
+				},
+			}
+
+			axiosGetStub.resolves(mockResponse)
+			const banners = await bannerService.fetchActiveBanners()
+
+			expect(banners).to.have.lengthOf(1)
+			expect(banners[0].id).to.equal("bnr_anthropic")
+		})
+
+		it("should NOT show banner when user has selected a different API provider", async () => {
+			const controllerWithAnthropic: Partial<Controller> = {
+				stateManager: {
+					getApiConfiguration: () => ({
+						actModeApiProvider: "anthropic",
+					}),
+					getGlobalSettingsKey: (key: string) => (key === "mode" ? "act" : undefined),
+					getGlobalStateKey: () => [],
+				} as any,
+			}
+			// Reinitialize with new controller
+			BannerService.reset()
+			bannerService = BannerService.initialize(controllerWithAnthropic as Controller)
 
 			const mockResponse = {
 				data: {
@@ -201,13 +239,13 @@ describe("BannerService", () => {
 			expect(banners).to.have.lengthOf(0)
 		})
 
-		it("should show banner if user has ANY of multiple specified providers", async () => {
+		it("should show banner if user has selected ANY of multiple specified providers", async () => {
 			const controllerWithAnthropic: Partial<Controller> = {
 				stateManager: {
 					getApiConfiguration: () => ({
-						apiKey: "sk-ant-test", // Has Anthropic key
+						actModeApiProvider: "anthropic",
 					}),
-					getGlobalSettingsKey: () => undefined,
+					getGlobalSettingsKey: (key: string) => (key === "mode" ? "act" : undefined),
 					getGlobalStateKey: () => [],
 				} as any,
 			}
@@ -237,6 +275,41 @@ describe("BannerService", () => {
 
 			expect(banners).to.have.lengthOf(1)
 			expect(banners[0].id).to.equal("bnr_multi")
+		})
+
+		it("should NOT show banner when no provider is selected", async () => {
+			const controllerWithNoProvider: Partial<Controller> = {
+				stateManager: {
+					getApiConfiguration: () => ({}),
+					getGlobalSettingsKey: (key: string) => (key === "mode" ? "act" : undefined),
+					getGlobalStateKey: () => [],
+				} as any,
+			}
+			// Reinitialize with new controller
+			BannerService.reset()
+			bannerService = BannerService.initialize(controllerWithNoProvider as Controller)
+
+			const mockResponse = {
+				data: {
+					data: {
+						items: [
+							{
+								id: "bnr_openai",
+								titleMd: "OpenAI Users",
+								bodyMd: "For OpenAI API",
+								severity: "info" as const,
+								placement: "top" as const,
+								rulesJson: JSON.stringify({ providers: ["openai"] } as BannerRules),
+							},
+						],
+					},
+				},
+			}
+
+			axiosGetStub.resolves(mockResponse)
+			const banners = await bannerService.fetchActiveBanners()
+
+			expect(banners).to.have.lengthOf(0)
 		})
 	})
 
