@@ -1,11 +1,11 @@
 import { PostHog } from "posthog-node"
-import * as vscode from "vscode"
 import { HostProvider } from "@/hosts/host-provider"
 import { getDistinctId } from "@/services/logging/distinctId"
 import { PostHogClientProvider } from "@/services/telemetry/providers/posthog/PostHogClientProvider"
 import { Setting } from "@/shared/proto/index.host"
 import * as pkg from "../../../../package.json"
 import { PostHogClientValidConfig } from "../../../shared/services/config/posthog-config"
+import { getErrorLevelFromString } from ".."
 import { ClineError } from "../ClineError"
 import type { ErrorSettings, IErrorProvider } from "./IErrorProvider"
 
@@ -53,13 +53,8 @@ export class PostHogErrorProvider implements IErrorProvider {
 			this.errorSettings.hostEnabled = false
 		}
 
-		// Check extension-specific telemetry setting
-		const config = vscode.workspace.getConfiguration("cline")
-		if (config.get("telemetrySetting") === "disabled") {
-			this.errorSettings.enabled = false
-		}
+		this.errorSettings.level = getErrorLevelFromString(hostSettings.errorLevel)
 
-		this.errorSettings.level = await this.getErrorLevel()
 		return this
 	}
 
@@ -132,15 +127,6 @@ export class PostHogErrorProvider implements IErrorProvider {
 
 	public getSettings(): ErrorSettings {
 		return { ...this.errorSettings }
-	}
-
-	private async getErrorLevel(): Promise<ErrorSettings["level"]> {
-		const hostSettings = await HostProvider.env.getTelemetrySettings({})
-		if (hostSettings.isEnabled === Setting.DISABLED) {
-			return "off"
-		}
-		const config = vscode.workspace.getConfiguration("telemetry")
-		return config?.get<ErrorSettings["level"]>("telemetryLevel") || "all"
 	}
 
 	private get distinctId(): string {
