@@ -8,8 +8,14 @@ export {
 	TelemetryProviderFactory,
 	type TelemetryProviderType,
 } from "./TelemetryProviderFactory"
-
-// Export the enums for terminal telemetry
+// Export terminal type definitions for type-safe telemetry
+export type {
+	StandaloneOutputMethod,
+	TerminalOutputMethod,
+	TerminalType,
+	VscodeOutputMethod,
+} from "./TelemetryService"
+// Export the enums and types for terminal telemetry
 export {
 	TerminalHangStage,
 	TerminalOutputFailureReason,
