@@ -59,25 +59,25 @@ e2e("Views - can set up API keys and navigate to Settings from Chat", async ({ s
 
 	// Verify What's New Section is showing and starts with first banner,
 	// and the navigation buttons work
-	await expect(sidebar.locator(".fade-in-cards")).toBeVisible()
+	await expect(sidebar.locator(".animate-fade-in")).toBeVisible()
 	await expect(
 		sidebar
 			.locator("div")
-			.filter({ hasText: /^1\/3$/ })
+			.filter({ hasText: /^1 \/ 3$/ })
 			.first(),
 	).toBeVisible()
 	await sidebar.getByRole("button", { name: "Next banner" }).click()
 	await expect(
 		sidebar
 			.locator("div")
-			.filter({ hasText: /^2\/3$/ })
+			.filter({ hasText: /^2 \/ 3$/ })
 			.first(),
 	).toBeVisible()
 	await sidebar.getByRole("button", { name: "Previous banner" }).click()
 	await expect(
 		sidebar
 			.locator("div")
-			.filter({ hasText: /^1\/3$/ })
+			.filter({ hasText: /^1 \/ 3$/ })
 			.first(),
 	).toBeVisible()
 })
