@@ -1,8 +1,7 @@
+import { BANNER_DATA, BannerAction, BannerActionType, BannerCardData } from "@shared/cline/banner"
 import { EmptyRequest, Int64Request } from "@shared/proto/index.cline"
-import { VSCodeLink } from "@vscode/webview-ui-toolkit/react"
-import { Megaphone, Terminal } from "lucide-react"
 import React, { useCallback, useEffect, useMemo, useState } from "react"
-import BannerCarousel, { BannerData } from "@/components/common/BannerCarousel"
+import BannerCarousel from "@/components/common/BannerCarousel"
 import { CURRENT_CLI_BANNER_VERSION } from "@/components/common/CliInstallBanner"
 import { CURRENT_INFO_BANNER_VERSION } from "@/components/common/InfoBanner"
 import { CURRENT_MODEL_BANNER_VERSION } from "@/components/common/NewModelBanner"
@@ -11,11 +10,11 @@ import HistoryPreview from "@/components/history/HistoryPreview"
 import { useApiConfigurationHandlers } from "@/components/settings/utils/useApiConfigurationHandlers"
 import HomeHeader from "@/components/welcome/HomeHeader"
 import { SuggestedTasks } from "@/components/welcome/SuggestedTasks"
-import { PLATFORM_CONFIG, PlatformType } from "@/config/platform.config"
 import { useClineAuth } from "@/context/ClineAuthContext"
 import { useExtensionState } from "@/context/ExtensionStateContext"
 import { AccountServiceClient, StateServiceClient } from "@/services/grpc-client"
-import { isMacOSOrLinux } from "@/utils/platformUtils"
+import { convertBannerData } from "@/utils/bannerUtils"
+import { getCurrentPlatform } from "@/utils/platformUtils"
 import { WelcomeSectionProps } from "../../types/chatTypes"
 
 /**
@@ -36,15 +35,6 @@ export const WelcomeSection: React.FC<WelcomeSectionProps> = ({
 	const [hasShownWhatsNewModal, setHasShownWhatsNewModal] = useState(false)
 	const [showWhatsNewModal, setShowWhatsNewModal] = useState(false)
 
-	const shouldShowInfoBanner = lastDismissedInfoBannerVersion < CURRENT_INFO_BANNER_VERSION
-	const shouldShowNewModelBanner = lastDismissedModelBannerVersion < CURRENT_MODEL_BANNER_VERSION
-
-	// Show CLI banner if not dismissed and platform is VSCode (not JetBrains/standalone)
-	const shouldShowCliBanner =
-		isMacOSOrLinux() &&
-		PLATFORM_CONFIG.type === PlatformType.VSCODE &&
-		lastDismissedCliBannerVersion < CURRENT_CLI_BANNER_VERSION
-
 	const { clineUser } = useClineAuth()
 	const { openRouterModels, setShowChatModelSelector, navigateToSettings, subagentsEnabled } = useExtensionState()
 	const { handleFieldsChange } = useApiConfigurationHandlers()
@@ -63,176 +53,143 @@ export const WelcomeSection: React.FC<WelcomeSectionProps> = ({
 		hideAnnouncement()
 	}, [hideAnnouncement])
 
-	// Build array of active banners for carousel
-	const activeBanners = useMemo((): BannerData[] => {
-		const banners: BannerData[] = []
+	/**
+	 * Check if a banner has been dismissed based on its version
+	 */
+	const isBannerDismissed = useCallback(
+		(bannerId: string): boolean => {
+			if (bannerId.startsWith("info-banner")) {
+				return (lastDismissedInfoBannerVersion ?? 0) >= CURRENT_INFO_BANNER_VERSION
+			}
+			if (bannerId.startsWith("new-model")) {
+				return (lastDismissedModelBannerVersion ?? 0) >= CURRENT_MODEL_BANNER_VERSION
+			}
+			if (bannerId.startsWith("cli-")) {
+				return (lastDismissedCliBannerVersion ?? 0) >= CURRENT_CLI_BANNER_VERSION
+			}
+			return false
+		},
+		[lastDismissedInfoBannerVersion, lastDismissedModelBannerVersion, lastDismissedCliBannerVersion],
+	)
 
-		if (shouldShowInfoBanner) {
-			banners.push({
-				id: "info-banner",
-				icon: <span>💡</span>,
-				title: "Use Cline in Right Sidebar",
-				description: (
-					<>
-						For the best experience, drag the Cline icon to your right sidebar. This keeps your file explorer and
-						editor visible while you chat with Cline, making it easier to navigate your codebase and see changes in
-						real-time.{" "}
-						<VSCodeLink
-							className="cursor-pointer"
-							href="https://docs.cline.bot/features/customization/opening-cline-in-sidebar"
-							style={{ display: "inline" }}>
-							See how →
-						</VSCodeLink>
-					</>
-				),
-				onDismiss: () => {
-					StateServiceClient.updateInfoBannerVersion({ value: CURRENT_INFO_BANNER_VERSION }).catch(console.error)
-				},
-			})
-		}
-
-		if (shouldShowNewModelBanner) {
-			const setNewModel = () => {
-				const modelId = "anthropic/claude-opus-4.5"
-				handleFieldsChange({
-					planModeOpenRouterModelId: modelId,
-					actModeOpenRouterModelId: modelId,
-					planModeOpenRouterModelInfo: openRouterModels[modelId],
-					actModeOpenRouterModelInfo: openRouterModels[modelId],
-					planModeApiProvider: "cline",
-					actModeApiProvider: "cline",
-				})
-				setTimeout(() => setShowChatModelSelector(true), 10)
+	/**
+	 * Banner configuration from backend
+	 * In production, this would come from an API/gRPC call
+	 * For now, using EXAMPLE_BANNER_DATA with version-based filtering
+	 */
+	const bannerConfig = useMemo((): BannerCardData[] => {
+		// Filter banners based on version tracking and user status
+		return BANNER_DATA.filter((banner) => {
+			if (isBannerDismissed(banner.id)) {
+				return false
 			}
 
-			const handleShowAccount = () => {
-				AccountServiceClient.accountLoginClicked(EmptyRequest.create()).catch((err) =>
-					console.error("Failed to get login URL:", err),
-				)
+			if (banner.isClineUserOnly !== undefined) {
+				return banner.isClineUserOnly === !!clineUser
 			}
 
-			banners.push({
-				id: "new-model",
-				icon: <Megaphone className="w-5 h-5" />,
-				title: "Claude Opus 4.5 Now Available",
-				description: "State-of-the-art performance at 3x lower cost than Opus 4.1. Available now in the Cline provider.",
-				actions: [
-					{
-						label: clineUser ? "Try Now" : "Get Started",
-						onClick: clineUser ? setNewModel : handleShowAccount,
-						variant: "primary",
-					},
-				],
-				onDismiss: () => {
-					StateServiceClient.updateModelBannerVersion(
-						Int64Request.create({ value: CURRENT_MODEL_BANNER_VERSION }),
-					).catch(console.error)
-				},
-			})
-		}
+			if (banner.platforms && !banner.platforms.includes(getCurrentPlatform())) {
+				return false
+			}
 
-		if (shouldShowCliBanner) {
-			const handleInstallCli = async () => {
-				try {
-					await StateServiceClient.installClineCli(EmptyRequest.create())
-				} catch (error) {
-					console.error("Failed to initiate CLI installation:", error)
+			return true
+		})
+	}, [isBannerDismissed, clineUser])
+
+	/**
+	 * Action handler - maps action types to actual implementations
+	 */
+	const handleBannerAction = useCallback(
+		(action: BannerAction) => {
+			switch (action.action) {
+				case BannerActionType.Link:
+					// Links are handled by VSCodeLink component
+					break
+
+				case BannerActionType.SetModel: {
+					const modelId = action.arg || "anthropic/claude-opus-4.5"
+					handleFieldsChange({
+						planModeOpenRouterModelId: modelId,
+						actModeOpenRouterModelId: modelId,
+						planModeOpenRouterModelInfo: openRouterModels[modelId],
+						actModeOpenRouterModelInfo: openRouterModels[modelId],
+						planModeApiProvider: "cline",
+						actModeApiProvider: "cline",
+					})
+					setTimeout(() => setShowChatModelSelector(true), 10)
+					break
 				}
-			}
 
-			const handleEnableSubagents = () => {
-				if (!subagentsEnabled) {
-					navigateToSettings("features")
-				}
-			}
-
-			banners.push({
-				id: "cli-install",
-				icon: <Terminal className="w-5 h-5" />,
-				title: isMacOSOrLinux() ? "CLI & Subagents Available" : "Cline CLI Info",
-				description: isMacOSOrLinux() ? (
-					<>
-						Use Cline in your terminal and enable subagent capabilities.{" "}
-						<VSCodeLink href="https://docs.cline.bot/cline-cli/overview" style={{ display: "inline" }}>
-							Learn more
-						</VSCodeLink>
-					</>
-				) : (
-					<>
-						Available for macOS and Linux. Coming soon to other platforms.{" "}
-						<VSCodeLink href="https://docs.cline.bot/cline-cli/overview" style={{ display: "inline" }}>
-							Learn more
-						</VSCodeLink>
-					</>
-				),
-				actions: isMacOSOrLinux()
-					? [
-							{ label: "Install", onClick: handleInstallCli, variant: "primary" },
-							{
-								label: "Enable Subagents",
-								onClick: handleEnableSubagents,
-								variant: "primary",
-								disabled: subagentsEnabled,
-							},
-						]
-					: [
-							{ label: "Install CLI", onClick: handleInstallCli, variant: "primary" },
-							{ label: "Subagents (Windows coming soon)", onClick: () => {}, variant: "secondary", disabled: true },
-						],
-				onDismiss: () => {
-					StateServiceClient.updateCliBannerVersion(Int64Request.create({ value: CURRENT_CLI_BANNER_VERSION })).catch(
-						console.error,
+				case BannerActionType.ShowAccount:
+					AccountServiceClient.accountLoginClicked(EmptyRequest.create()).catch((err) =>
+						console.error("Failed to get login URL:", err),
 					)
-				},
-			})
-		}
+					break
 
-		return banners
-	}, [
-		shouldShowInfoBanner,
-		shouldShowNewModelBanner,
-		shouldShowCliBanner,
-		clineUser,
-		openRouterModels,
-		setShowChatModelSelector,
-		handleFieldsChange,
-		navigateToSettings,
-		subagentsEnabled,
-	])
+				case BannerActionType.ShowApiSettings:
+					navigateToSettings("api")
+					break
+
+				case BannerActionType.ShowFeatureSettings:
+					navigateToSettings("features")
+					break
+
+				case BannerActionType.InstallCli:
+					StateServiceClient.installClineCli(EmptyRequest.create()).catch((error) =>
+						console.error("Failed to initiate CLI installation:", error),
+					)
+					break
+
+				default:
+					console.warn("Unknown banner action:", action.action)
+			}
+		},
+		[handleFieldsChange, openRouterModels, setShowChatModelSelector, navigateToSettings],
+	)
+
+	/**
+	 * Dismissal handler - updates version tracking
+	 */
+	const handleBannerDismiss = useCallback((bannerId: string) => {
+		// Map banner IDs to version updates
+		if (bannerId.startsWith("info-banner")) {
+			StateServiceClient.updateInfoBannerVersion({ value: CURRENT_INFO_BANNER_VERSION }).catch(console.error)
+		} else if (bannerId.startsWith("new-model")) {
+			StateServiceClient.updateModelBannerVersion(Int64Request.create({ value: CURRENT_MODEL_BANNER_VERSION })).catch(
+				console.error,
+			)
+		} else if (bannerId.startsWith("cli-")) {
+			StateServiceClient.updateCliBannerVersion(Int64Request.create({ value: CURRENT_CLI_BANNER_VERSION })).catch(
+				console.error,
+			)
+		}
+	}, [])
+
+	/**
+	 * Build array of active banners for carousel
+	 */
+	const activeBanners = useMemo(() => {
+		// Convert to BannerData format for carousel
+		return bannerConfig.map((banner) =>
+			convertBannerData(banner, {
+				onAction: handleBannerAction,
+				onDismiss: handleBannerDismiss,
+			}),
+		)
+	}, [bannerConfig, clineUser, subagentsEnabled, handleBannerAction, handleBannerDismiss])
 
 	return (
 		<div className="flex flex-col flex-1 w-full h-full p-0 m-0">
-			<style>
-				{`
-					@keyframes fadeIn {
-						from {
-							opacity: 0;
-							transform: scale(0.98);
-						}
-						to {
-							opacity: 1;
-							transform: scale(1);
-						}
-					}
-					.fade-in-cards {
-						animation: fadeIn 0.4s ease-out forwards;
-					}
-					.fade-in-history {
-						animation: fadeIn 0.4s ease-out forwards;
-						opacity: 0;
-					}
-				`}
-			</style>
 			<WhatsNewModal onClose={handleCloseWhatsNewModal} open={showWhatsNewModal} version={version} />
 			<div className="overflow-y-auto flex flex-col pb-2.5">
 				<HomeHeader shouldShowQuickWins={shouldShowQuickWins} />
 				{!showWhatsNewModal && (
 					<>
-						<div className="fade-in-cards">
+						<div className="animate-fade-in">
 							<BannerCarousel banners={activeBanners} />
 						</div>
 						{!shouldShowQuickWins && taskHistory.length > 0 && (
-							<div className="fade-in-history">
+							<div className="animate-fade-in opacity-0">
 								<HistoryPreview showHistoryView={showHistoryView} />
 							</div>
 						)}
