@@ -167,7 +167,11 @@ const ServerRow = ({
 
 	return (
 		<div className="mb-2.5">
-			<div className="flex bg-code-block-background p-2 gap-4 items-center" onClick={handleRowClick}>
+			<div
+				className={cn("flex bg-code-block-background p-2 gap-4 items-center", {
+					"cursor-pointer": !server.error && isExpandable,
+				})}
+				onClick={handleRowClick}>
 				{!server.error && isExpandable && (
 					<span
 						className={cn("mr-2 codicon", {
@@ -207,7 +211,14 @@ const ServerRow = ({
 					</Button>
 				)}
 				{/* Toggle Switch */}
-				<Switch checked={!server.disabled} key={server.name} onClick={handleToggleMcpServer} />
+				<Switch
+					checked={!server.disabled}
+					key={server.name}
+					onClick={(e) => {
+						e.stopPropagation()
+						handleToggleMcpServer()
+					}}
+				/>
 				<div
 					className={cn("h-2 w-2 ml-0.5 rounded-full", {
 						"bg-success": server.status === "connected",
