@@ -42,10 +42,25 @@ export const isChrome = userAgent.indexOf("Chrome") >= 0
 export const isSafari = !isChrome && userAgent.indexOf("Safari") >= 0
 
 /**
+ * Gets the current platform: 'windows', 'mac', or 'linux'
+ * Defaults to 'linux' if platform cannot be determined
+ */
+export function getCurrentPlatform() {
+	// Fallback to linux if platform is not available
+	switch (process?.platform) {
+		case "win32":
+			return "windows"
+		case "darwin":
+			return "mac"
+		default:
+			return "linux"
+	}
+}
+
+/**
  * Checks if the platform is macOS or Linux
  * @returns true if platform is darwin (macOS) or linux
  */
 export const isMacOSOrLinux = (): boolean => {
-	const platform = process?.platform
-	return !platform?.startsWith("win") // Non-Windows
+	return getCurrentPlatform() !== "windows" // Non-Windows
 }
